refactor: move func
