+++ src/pivot-table/hooks/use-column-width/use-column-width.ts
@@ -1,8 +1,12 @@
-import { useEffect, useMemo } from "react";
+import { useMeasureText } from "@qlik/nebula-table-utils/lib/hooks";
+import { useCallback, useEffect, useMemo } from "react";
 import { GRID_BORDER } from "../../constants";
-import type { ColumnWidthHook } from "./types";
+import { useStyleContext } from "../../contexts/StyleProvider";
+import { LOCK_ICON_SIZE, MENU_ICON_SIZE, TOTAL_CELL_PADDING } from "./constants";
+import type { ColumnWidthHook, GetHeaderCellsIconsVisibilityStatus } from "./types";
 import useColumnWidthLeft from "./use-column-width-left";
 import useColumnWidthRight from "./use-column-width-right";
+import { getMeasureTextArgs } from "./utils";
 
 export default function useColumnWidth({
   layoutService,
@@ -12,6 +16,9 @@ export default function useColumnWidth({
   verticalScrollbarWidth,
   horizontalScrollbarHeightSetter,
 }: ColumnWidthHook) {
+  const styleService = useStyleContext();
+  const { measureText: measureTextForHeader } = useMeasureText(getMeasureTextArgs(styleService.header));
+
   const leftWidths = useColumnWidthLeft({ layoutService, tableRect, headersData });
 
   const rightWidths = useColumnWidthRight({
@@ -29,6 +36,31 @@ export default function useColumnWidth({
   );
 
   const showLastRightBorder = useMemo(() => totalWidth < tableRect.width, [totalWidth, tableRect.width]);
+
+  const getHeaderCellsIconsVisibilityStatus = useCallback<GetHeaderCellsIconsVisibilityStatus>(
+    (idx, isLocked, title = "") => {
+      const colWidth = leftWidths.leftGridColumnWidths[idx];
+      let shouldShowMenuIcon = false;
+      let shouldShowLockIcon = false;
+      const measuredTextForHeader = measureTextForHeader(title);
+
+      let headerSize = measuredTextForHeader + TOTAL_CELL_PADDING;
+
+      if (isLocked && headerSize + LOCK_ICON_SIZE <= colWidth) {
+        shouldShowLockIcon = true;
+        headerSize += LOCK_ICON_SIZE;
+      }
+      if (headerSize + MENU_ICON_SIZE <= colWidth) {
+        shouldShowMenuIcon = true;
+      }
+
+      return {
+        shouldShowMenuIcon,
+        shouldShowLockIcon,
+      };
+    },
+    [leftWidths, measureTextForHeader],
+  );
 
   // Horizontal scrollbar height control based on columns (full) visibility
   useEffect(() => {
@@ -49,5 +81,6 @@ export default function useColumnWidth({
     ...rightWidths,
     totalWidth,
     showLastRightBorder,
+    getHeaderCellsIconsVisibilityStatus,
   };
 }
